move search index build into re_dev_tools
