+++ Cargo.toml
@@ -30,7 +30,6 @@ version = "0.16.0-alpha.1+dev"
 # re_log_types 0.3.0-alpha.0, NOT 0.3.0-alpha.4 even though it is newer and semver-compatible.
 re_analytics = { path = "crates/re_analytics", version = "=0.16.0-alpha.1", default-features = false }
 re_build_info = { path = "crates/re_build_info", version = "=0.16.0-alpha.1", default-features = false }
-re_build_search_index = { path = "crates/re_build_search_index", version = "=0.16.0-alpha.1", default-features = false }
 re_build_tools = { path = "crates/re_build_tools", version = "=0.16.0-alpha.1", default-features = false }
 re_build_web_viewer = { path = "crates/re_build_web_viewer", version = "=0.16.0-alpha.1", default-features = false }
 re_crash_handler = { path = "crates/re_crash_handler", version = "=0.16.0-alpha.1", default-features = false }
