--- crates/re_build_search_index/Cargo.toml
@@ -1,39 +1,0 @@
-[package]
-name = "re_build_search_index"
-authors.workspace = true
-description = "Build a Meilisearch index from all of Rerun's documentation"
-edition.workspace = true
-homepage.workspace = true
-include.workspace = true
-license.workspace = true
-publish = false
-readme = "README.md"
-repository.workspace = true
-rust-version.workspace = true
-version.workspace = true
-
-[package.metadata.docs.rs]
-all-features = true
-
-
-[dependencies]
-re_build_tools.workspace = true
-
-# External
-anyhow.workspace = true
-argh.workspace = true
-camino.workspace = true
-cargo_metadata.workspace = true
-glob.workspace = true
-indicatif.workspace = true
-itertools.workspace = true
-roxmltree = "0.19.0"
-rayon.workspace = true
-rustdoc-json = "0.8.8"
-rustdoc-types = "0.24.0"
-serde = { workspace = true, features = ["derive"] }
-serde_json.workspace = true
-serde_yaml.workspace = true
-toml = { workspace = true, features = ["parse", "preserve_order"] }
-ureq = { workspace = true, features = ["json"] }
-url.workspace = true
