+++ crates/re_dev_tools/src/build_search_index/build.rs
@@ -1,4 +1,4 @@
-use crate::{ingest, meili, DEFAULT_INDEX, DEFAULT_KEY, DEFAULT_URL};
+use super::{ingest, meili, DEFAULT_INDEX, DEFAULT_KEY, DEFAULT_URL};
 
 /// Index documentation, examples, and API references for all languages
 #[derive(argh::FromArgs)]
