+++ crates/re_dev_tools/src/build_search_index/ingest/rust.rs
