+++ crates/re_dev_tools/src/build_search_index/mod.rs
@@ -7,30 +7,12 @@ mod meili;
 mod repl;
 mod util;
 
-fn main() -> anyhow::Result<()> {
-    let args: Args = argh::from_env();
-
-    let Some(cmd) = args.cmd else {
-        let help_text = <Args as argh::FromArgs>::from_args(&["search-index"], &["--help"])
-            .err()
-            .unwrap()
-            .output;
-
-        eprintln!("{help_text}");
-        return Ok(());
-    };
-
-    match cmd {
-        Cmd::Repl(cmd) => cmd.run(),
-        Cmd::Build(cmd) => cmd.run(),
-    }
-}
-
 /// Meilisearch indexer and REPL
 #[derive(argh::FromArgs)]
-struct Args {
+#[argh(subcommand, name = "search-index")]
+pub struct Args {
     #[argh(subcommand)]
-    cmd: Option<Cmd>,
+    cmd: Cmd,
 }
 
 #[derive(argh::FromArgs)]
@@ -38,6 +20,13 @@ struct Args {
 enum Cmd {
     Repl(repl::Repl),
     Build(build::Build),
+}
+
+pub fn main(args: Args) -> anyhow::Result<()> {
+    match args.cmd {
+        Cmd::Repl(cmd) => cmd.run(),
+        Cmd::Build(cmd) => cmd.run(),
+    }
 }
 
 const DEFAULT_URL: &str = "http://localhost:7700";
