+++ pixi.toml
@@ -82,8 +82,6 @@ rerun-build-web-release = "rustup target add wasm32-unknown-unknown && cargo run
 
 rs-check = "rustup target add wasm32-unknown-unknown && python scripts/ci/rust_checks.py"
 
-build-examples = "cargo run -q --locked -p re_dev_tools -- build-examples"
-
 # Code formatting for all languages.
 format = { depends_on = ["py-fmt", "cpp-fmt", "toml-fmt", "misc-fmt"] }
 fmt = { depends_on = ["format"] }
@@ -137,9 +135,12 @@ py-plot-dashboard = { cmd = "python tests/python/plot_dashboard_stress/main.py",
 ] }
 rs-plot-dashboard = { cmd = "cargo r -p plot_dashboard_stress --release --" }
 
+dev-tools = "cargo run -q --locked -p re_dev_tools --"
+build-examples = "pixi run dev-tools build-examples"
+
 # Build the documentation search index.
 # See `pixi run search-index --help` for more information.
-search-index = "cargo run -p re_build_search_index --release --"
+search-index = "pixi run dev-tools search-index"
 
 # Start a local meilisearch instance at `localhost:7700` with master key `test`.
 # This should only be used for testing the search index locally.
