+++ rerun_cpp/docs/Doxyfile
@@ -1773,7 +1773,7 @@ GENERATE_MAN           = NO
 # captures the structure of the code including all documentation.
 # The default value is: NO.
 
-GENERATE_XML           = YES  # Generated files are used in `re_build_search_index`
+GENERATE_XML           = YES  # Generated files are used in `build_search_index`
 
 #---------------------------------------------------------------------------
 # Configuration options related to the DOCBOOK output
